RA-07 Внесены правки в макет, классы данных вынесены в отдельную папку
