+++ app/src/main/java/com/example/recipesapp/MainActivity.kt
@@ -7,13 +7,13 @@ import androidx.fragment.app.commit
 import androidx.fragment.app.replace
 import com.example.recipesapp.databinding.ActivityMainBinding
 
-private var _binding: ActivityMainBinding? = null
-private val binding
-    get() = _binding
-        ?: throw IllegalStateException("Binding для ActivityMainBinding не должен быть null")
-
-
 class MainActivity : AppCompatActivity() {
+
+    private var _binding: ActivityMainBinding? = null
+    private val binding
+        get() = _binding
+            ?: throw IllegalStateException("Binding для ActivityMainBinding не должен быть null")
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         _binding = ActivityMainBinding.inflate(layoutInflater)
