--- app/src/main/java/com/example/recipesapp/MealsData.kt
@@ -1,22 +1,0 @@
-package com.example.recipesapp
-
-data class Category(
-    val id: Int,
-    val title: String,
-    val description: String,
-    val imageUrl: String,
-)
-
-data class Recipe(
-    val id: Int,
-    val title: String,
-    val ingredients: List<Ingredient>,
-    val method: String,
-    val imageUrl: String,
-)
-
-data class Ingredient(
-    val quantity: String,
-    val unitOfMeasure: String,
-    val description: String,
-)
