+++ app/src/main/java/entity/Category.kt
@@ -1,0 +1,8 @@
+package entity
+
+data class Category(
+    val id: Int,
+    val title: String,
+    val description: String,
+    val imageUrl: String,
+)
