+++ app/src/main/java/entity/Ingredient.kt
@@ -1,0 +1,7 @@
+package entity
+
+data class Ingredient(
+    val quantity: String,
+    val unitOfMeasure: String,
+    val description: String,
+)
