+++ app/src/main/java/entity/Recipe.kt
@@ -1,0 +1,9 @@
+package entity
+
+data class Recipe(
+    val id: Int,
+    val title: String,
+    val ingredients: List<Ingredient>,
+    val method: String,
+    val imageUrl: String,
+)
