+++ app/src/main/res/layout/item_category.xml
@@ -1,12 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="@dimen/card_width"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
     android:layout_height="@dimen/card_height"
-    android:clickable="true"
-    app:cardCornerRadius="@dimen/space_half"
     android:layout_marginHorizontal="@dimen/space_half"
-    android:layout_marginVertical="@dimen/space_half">
+    android:layout_marginVertical="@dimen/space_half"
+    android:clickable="true"
+    app:cardCornerRadius="@dimen/space_half">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -17,23 +18,22 @@
             android:layout_width="match_parent"
             android:layout_height="@dimen/image_size_medium"
             android:scaleType="centerCrop"
-            android:src="@drawable/bcg_categories" />
+            tools:src="@drawable/bcg_categories" />
 
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginVertical="@dimen/space_half"
-            android:layout_marginStart="@dimen/space_half"
-            android:text="ЗАГОЛОВОК"
-            android:textAppearance="@style/StyleTitle14Secondary" />
+            android:layout_margin="@dimen/space_half"
+            android:textAppearance="@style/StyleTitle14Secondary"
+            tools:text="ЗАГОЛОВОК" />
 
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginHorizontal="@dimen/space_half"
             android:layout_marginBottom="@dimen/space_half"
-            android:text="Описание"
-            android:textAppearance="@style/StyleMainText" />
+            android:textAppearance="@style/StyleMainText"
+            tools:text="Описание" />
 
     </LinearLayout>
